OCM-8418 | test: apply the automation labels on testing
update

update

update

update

update

update
