+++ tests/ci/labels/category.go
@@ -1,0 +1,8 @@
+package labels
+
+import (
+	. "github.com/onsi/ginkgo/v2"
+)
+
+// Exclude in CI
+var Exclude = Label("Exclude")
