+++ tests/e2e/test_rosacli_cluster.go
@@ -21,8 +21,7 @@ import (
 )
 
 var _ = Describe("Edit cluster",
-	labels.Day2,
-	labels.FeatureCluster,
+	labels.Feature.Cluster,
 	func() {
 		defer GinkgoRecover()
 
@@ -54,7 +53,7 @@ var _ = Describe("Edit cluster",
 		})
 
 		It("can check the description of the cluster - [id:34102]",
-			labels.Medium,
+			labels.Medium, labels.Runtime.Day2,
 			func() {
 				By("Describe cluster in text format")
 				output, err := clusterService.DescribeCluster(clusterID)
@@ -109,13 +108,13 @@ var _ = Describe("Edit cluster",
 			})
 
 		It("can restrict master API endpoint to direct, private connectivity or not - [id:38850]",
-			labels.High,
+			labels.High, labels.Runtime.Day2,
 			func() {
 				By("Check the cluster is not private cluster")
 				private, err := clusterService.IsPrivateCluster(clusterID)
 				Expect(err).To(BeNil())
 				if private {
-					Skip("This case needs to test on private cluster as the prerequirement,it was not fullfilled, skip the case!!")
+					SkipTestOnFeature("private")
 				}
 				isSTS, err := clusterService.IsSTSCluster(clusterID)
 				Expect(err).To(BeNil())
@@ -169,7 +168,7 @@ var _ = Describe("Edit cluster",
 
 		// OCM-5231 caused the description parser issue
 		It("can disable workload monitoring on/off - [id:45159]",
-			labels.High,
+			labels.High, labels.Runtime.Day2,
 			func() {
 				By("Check the cluster UWM is in expected status")
 				output, err := clusterService.DescribeCluster(clusterID)
@@ -220,7 +219,7 @@ var _ = Describe("Edit cluster",
 			})
 
 		It("can validate for deletion of upgrade policy of rosa cluster - [id:38787]",
-			labels.Medium,
+			labels.Medium, labels.Runtime.Day2,
 			func() {
 				By("Validate that deletion of upgrade policy for rosa cluster will work via rosacli")
 				output, err := clusterService.DeleteUpgrade()
@@ -239,16 +238,22 @@ var _ = Describe("Edit cluster",
 				Expect(err).To(HaveOccurred())
 				textData = rosaClient.Parser.TextData.Input(output).Parse().Tip()
 				Expect(textData).Should(ContainSubstring("Error: unknown flag: --interactive"))
-
 			})
 
 		It("validation for create/delete upgrade policies for hypershift clusters via rosacli should work well - [id:73814]",
-			labels.Medium, labels.NonClassicCluster,
+			labels.Medium, labels.Runtime.Day2,
 			func() {
 				defer func() {
 					_, err := clusterService.DeleteUpgrade("-c", clusterID, "-y")
 					Expect(err).ToNot(HaveOccurred())
 				}()
+
+				By("Skip testing if the cluster is not a HCP cluster")
+				hostedCluster, err := clusterService.IsHostedCPCluster(clusterID)
+				Expect(err).ToNot(HaveOccurred())
+				if !hostedCluster {
+					SkipNotHosted()
+				}
 
 				By("Upgrade cluster without --control-plane flag")
 				output, err := clusterService.Upgrade("-c", clusterID)
@@ -289,13 +294,13 @@ var _ = Describe("Edit cluster",
 			})
 
 		It("can allow sts cluster installation with compatible policies - [id:45161]",
-			labels.High,
+			labels.High, labels.Runtime.Day2,
 			func() {
 				By("Check the cluster is STS cluster or skip")
 				isSTSCluster, err := clusterService.IsSTSCluster(clusterID)
 				Expect(err).ToNot(HaveOccurred())
 				if !isSTSCluster {
-					Skip("This case 45161 is only supported on STS cluster")
+					SkipTestOnFeature("policy")
 				}
 
 				clusterName := "cluster-45161"
@@ -360,8 +365,7 @@ var _ = Describe("Edit cluster",
 	})
 
 var _ = Describe("Classic cluster creation validation",
-	labels.Day1Validation,
-	labels.FeatureCluster,
+	labels.Feature.Cluster,
 	func() {
 		defer GinkgoRecover()
 
@@ -391,200 +395,205 @@ var _ = Describe("Classic cluster creation validation",
 			Expect(len(errs)).To(Equal(0))
 		})
 
-		It("to check the basic validation for the classic rosa cluster creation by the rosa cli - [id:38770]", labels.Medium, labels.Day1Validation, labels.NonHCPCluster, func() {
-			var command string
-			var rosalCommand config.Command
-			By("Prepare creation command")
-			flags, err := profilehandler.GenerateClusterCreateFlags(profile, rosaClient)
-			Expect(err).To(BeNil())
+		It("to check the basic validation for the classic rosa cluster creation by the rosa cli - [id:38770]",
+			labels.Medium, labels.Runtime.Day1Negative,
+			func() {
+				By("Prepare creation command")
+				var command string
+				var rosalCommand config.Command
+				flags, err := profilehandler.GenerateClusterCreateFlags(profile, rosaClient)
+				Expect(err).To(BeNil())
 
-			command = "rosa create cluster --cluster-name " + profile.ClusterConfig.Name + " " + strings.Join(flags, " ")
-			rosalCommand = config.GenerateCommand(command)
+				command = "rosa create cluster --cluster-name " + profile.ClusterConfig.Name + " " + strings.Join(flags, " ")
+				rosalCommand = config.GenerateCommand(command)
 
-			rosalCommand.AddFlags("--dry-run")
-			originalClusterName := rosalCommand.GetFlagValue("--cluster-name", true)
-			originalMachineType := rosalCommand.GetFlagValue("--compute-machine-type", true)
-			originalRegion := rosalCommand.GetFlagValue("--region", true)
-			if !rosalCommand.CheckFlagExist("--replicas") {
-				rosalCommand.AddFlags("--replicas", "3")
-			}
-			originalReplicas := rosalCommand.GetFlagValue("--replicas", true)
-
-			invalidClusterNames := []string{
-				"1-test-1",
-				"-test-cluster",
-				"test-cluster-",
-			}
-			for _, cn := range invalidClusterNames {
-				By("Check the validation for cluster-name")
-				rosalCommand.ReplaceFlagValue(map[string]string{
-					"--cluster-name": cn,
-				})
-				stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-				Expect(err).NotTo(BeNil())
-				Expect(stdout.String()).To(ContainSubstring("Cluster name must consist of no more than 54 lowercase alphanumeric characters or '-', start with a letter, and end with an alphanumeric character"))
-			}
-
-			By("Check the validation for compute-machine-type")
-			invalidMachineType := "not-exist-machine-type"
-			rosalCommand.ReplaceFlagValue(map[string]string{
-				"--compute-machine-type": invalidMachineType,
-				"--cluster-name":         originalClusterName,
-			})
-			stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-			Expect(err).NotTo(BeNil())
-			Expect(stdout.String()).To(ContainSubstring("Expected a valid machine type"))
-
-			By("Check the validation for replicas")
-			invalidReplicasTypeErrorMap := map[string]string{
-				"4.5":  "invalid argument \"4.5\" for \"--replicas\" flag",
-				"five": "invalid argument \"five\" for \"--replicas\" flag",
-			}
-			for k, v := range invalidReplicasTypeErrorMap {
-				rosalCommand.ReplaceFlagValue(map[string]string{
-					"--compute-machine-type": originalMachineType,
-					"--replicas":             k,
-				})
-				stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-				Expect(err).NotTo(BeNil())
-				Expect(stdout.String()).To(ContainSubstring(v))
-			}
-			if rosalCommand.CheckFlagExist("--multi-az") {
-				invalidReplicasErrorMapMultiAZ := map[string]string{
-					"2":  "Multi AZ cluster requires at least 3 compute nodes",
-					"0":  "Multi AZ cluster requires at least 3 compute nodes",
-					"-3": "must be non-negative",
-					"5":  "Multi AZ clusters require that the number of compute nodes be a multiple of 3",
+				rosalCommand.AddFlags("--dry-run")
+				originalClusterName := rosalCommand.GetFlagValue("--cluster-name", true)
+				originalMachineType := rosalCommand.GetFlagValue("--compute-machine-type", true)
+				originalRegion := rosalCommand.GetFlagValue("--region", true)
+				if !rosalCommand.CheckFlagExist("--replicas") {
+					rosalCommand.AddFlags("--replicas", "3")
 				}
-				for k, v := range invalidReplicasErrorMapMultiAZ {
+				originalReplicas := rosalCommand.GetFlagValue("--replicas", true)
+
+				invalidClusterNames := []string{
+					"1-test-1",
+					"-test-cluster",
+					"test-cluster-",
+				}
+				for _, cn := range invalidClusterNames {
+					By("Check the validation for cluster-name")
 					rosalCommand.ReplaceFlagValue(map[string]string{
-						"--replicas": k,
+						"--cluster-name": cn,
+					})
+					stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+					Expect(err).NotTo(BeNil())
+					Expect(stdout.String()).To(ContainSubstring("Cluster name must consist of no more than 54 lowercase alphanumeric characters or '-', start with a letter, and end with an alphanumeric character"))
+				}
+
+				By("Check the validation for compute-machine-type")
+				invalidMachineType := "not-exist-machine-type"
+				rosalCommand.ReplaceFlagValue(map[string]string{
+					"--compute-machine-type": invalidMachineType,
+					"--cluster-name":         originalClusterName,
+				})
+				stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+				Expect(err).NotTo(BeNil())
+				Expect(stdout.String()).To(ContainSubstring("Expected a valid machine type"))
+
+				By("Check the validation for replicas")
+				invalidReplicasTypeErrorMap := map[string]string{
+					"4.5":  "invalid argument \"4.5\" for \"--replicas\" flag",
+					"five": "invalid argument \"five\" for \"--replicas\" flag",
+				}
+				for k, v := range invalidReplicasTypeErrorMap {
+					rosalCommand.ReplaceFlagValue(map[string]string{
+						"--compute-machine-type": originalMachineType,
+						"--replicas":             k,
 					})
 					stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
 					Expect(err).NotTo(BeNil())
 					Expect(stdout.String()).To(ContainSubstring(v))
 				}
-			} else {
-				invalidReplicasErrorMapSingeAZ := map[string]string{
-					"1":  "requires at least 2 compute nodes",
-					"0":  "requires at least 2 compute nodes",
-					"-1": "must be non-negative",
+				if rosalCommand.CheckFlagExist("--multi-az") {
+					invalidReplicasErrorMapMultiAZ := map[string]string{
+						"2":  "Multi AZ cluster requires at least 3 compute nodes",
+						"0":  "Multi AZ cluster requires at least 3 compute nodes",
+						"-3": "must be non-negative",
+						"5":  "Multi AZ clusters require that the number of compute nodes be a multiple of 3",
+					}
+					for k, v := range invalidReplicasErrorMapMultiAZ {
+						rosalCommand.ReplaceFlagValue(map[string]string{
+							"--replicas": k,
+						})
+						stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+						Expect(err).NotTo(BeNil())
+						Expect(stdout.String()).To(ContainSubstring(v))
+					}
+				} else {
+					invalidReplicasErrorMapSingeAZ := map[string]string{
+						"1":  "requires at least 2 compute nodes",
+						"0":  "requires at least 2 compute nodes",
+						"-1": "must be non-negative",
+					}
+					for k, v := range invalidReplicasErrorMapSingeAZ {
+						rosalCommand.ReplaceFlagValue(map[string]string{
+							"--replicas": k,
+						})
+						stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+						Expect(err).NotTo(BeNil())
+						Expect(stdout.String()).To(ContainSubstring(v))
+					}
 				}
-				for k, v := range invalidReplicasErrorMapSingeAZ {
-					rosalCommand.ReplaceFlagValue(map[string]string{
-						"--replicas": k,
-					})
-					stdout, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-					Expect(err).NotTo(BeNil())
-					Expect(stdout.String()).To(ContainSubstring(v))
-				}
-			}
 
-			By("Check the validation for region")
-			invalidRegion := "not-exist-region"
-			rosalCommand.ReplaceFlagValue(map[string]string{
-				"--region":   invalidRegion,
-				"--replicas": originalReplicas,
+				By("Check the validation for region")
+				invalidRegion := "not-exist-region"
+				rosalCommand.ReplaceFlagValue(map[string]string{
+					"--region":   invalidRegion,
+					"--replicas": originalReplicas,
+				})
+				stdout, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+				Expect(err).NotTo(BeNil())
+				Expect(stdout.String()).To(ContainSubstring("Unsupported region"))
+
+				By("Check the validation for billing-account for classic sts cluster")
+				rosalCommand.ReplaceFlagValue(map[string]string{
+					"--region": originalRegion,
+				})
+				rosalCommand.AddFlags("--billing-account", "123456789")
+				stdout, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+				Expect(err).NotTo(BeNil())
+				Expect(stdout.String()).To(ContainSubstring("Billing accounts are only supported for Hosted Control Plane clusters"))
 			})
-			stdout, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-			Expect(err).NotTo(BeNil())
-			Expect(stdout.String()).To(ContainSubstring("Unsupported region"))
 
-			By("Check the validation for billing-account for classic sts cluster")
-			rosalCommand.ReplaceFlagValue(map[string]string{
-				"--region": originalRegion,
+		It("to validate to create the sts cluster with invalid tag - [id:56440]",
+			labels.Medium, labels.Runtime.Day1Negative,
+			func() {
+				clusterService := rosaClient.Cluster
+				clusterName := "ocp-56440"
+
+				By("Create cluster with invalid tag key")
+				out, err := clusterService.CreateDryRun(
+					clusterName, "--tags=~~~:cluster",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("expected a valid user tag key '~~~' matching ^[\\pL\\pZ\\pN_.:/=+\\-@]{1,128}$"))
+
+				By("Create cluster with invalid tag value")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--tags=name:****",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("expected a valid user tag value '****' matching ^[\\pL\\pZ\\pN_.:/=+\\-@]{0,256}$"))
+
+				By("Create cluster with duplicate tag key")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--tags=name:test1,op:clound,name:test2",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("invalid tags, user tag keys must be unique, duplicate key 'name' found"))
+
+				By("Create cluster with invalid tag format")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--tags=test1,test2,test4",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("invalid tag format for tag '[test1]'. Expected tag format: 'key:value'"))
+
+				By("Create cluster with empty tag value")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--tags", "foo:",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("invalid tag format, tag key or tag value can not be empty"))
+
+				By("Create cluster with invalid tag format")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--tags=name:gender:age",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("invalid tag format for tag '[name gender age]'. Expected tag format: 'key:value'"))
+
 			})
-			rosalCommand.AddFlags("--billing-account", "123456789")
-			stdout, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-			Expect(err).NotTo(BeNil())
-			Expect(stdout.String()).To(ContainSubstring("Billing accounts are only supported for Hosted Control Plane clusters"))
-		})
 
-		It("to validate to create the sts cluster with invalid tag - [id:56440]", labels.Medium, labels.Day1Validation, func() {
-			clusterService := rosaClient.Cluster
-			clusterName := "ocp-56440"
+		It("to validate to create cluster with availability zones - [id:52692]",
+			labels.Medium, labels.Runtime.Day1Negative,
+			func() {
+				clusterService := rosaClient.Cluster
+				clusterName := "ocp-52692"
 
-			By("Create cluster with invalid tag key")
-			out, err := clusterService.CreateDryRun(
-				clusterName, "--tags=~~~:cluster",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("expected a valid user tag key '~~~' matching ^[\\pL\\pZ\\pN_.:/=+\\-@]{1,128}$"))
+				By("Create cluster with the zone not available in the region")
+				out, err := clusterService.CreateDryRun(
+					clusterName, "--availability-zones", "us-east-2e", "--region", "us-east-2",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("Expected a valid availability zone, 'us-east-2e' doesn't belong to region 'us-east-2' availability zones"))
 
-			By("Create cluster with invalid tag value")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--tags=name:****",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("expected a valid user tag value '****' matching ^[\\pL\\pZ\\pN_.:/=+\\-@]{0,256}$"))
+				By("Create cluster with zones not match region")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--availability-zones", "us-west-2b", "--region", "us-east-2",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("Expected a valid availability zone, 'us-west-2b' doesn't belong to region 'us-east-2' availability zones"))
 
-			By("Create cluster with duplicate tag key")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--tags=name:test1,op:clound,name:test2",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("invalid tags, user tag keys must be unique, duplicate key 'name' found"))
+				By("Create cluster with dup zones set")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--availability-zones", "us-west-2b,us-west-2b,us-west-2b", "--region", "us-west-2",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("Found duplicate Availability Zone: us-west-2b"))
 
-			By("Create cluster with invalid tag format")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--tags=test1,test2,test4",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("invalid tag format for tag '[test1]'. Expected tag format: 'key:value'"))
-
-			By("Create cluster with empty tag value")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--tags", "foo:",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("invalid tag format, tag key or tag value can not be empty"))
-
-			By("Create cluster with invalid tag format")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--tags=name:gender:age",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("invalid tag format for tag '[name gender age]'. Expected tag format: 'key:value'"))
-
-		})
-
-		It("to validate to create cluster with availability zones - [id:52692]", labels.Medium, labels.Day1Validation, func() {
-			clusterService := rosaClient.Cluster
-			clusterName := "ocp-52692"
-
-			By("Create cluster with the zone not available in the region")
-			out, err := clusterService.CreateDryRun(
-				clusterName, "--availability-zones", "us-east-2e", "--region", "us-east-2",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("Expected a valid availability zone, 'us-east-2e' doesn't belong to region 'us-east-2' availability zones"))
-
-			By("Create cluster with zones not match region")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--availability-zones", "us-west-2b", "--region", "us-east-2",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("Expected a valid availability zone, 'us-west-2b' doesn't belong to region 'us-east-2' availability zones"))
-
-			By("Create cluster with dup zones set")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--availability-zones", "us-west-2b,us-west-2b,us-west-2b", "--region", "us-west-2",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("Found duplicate Availability Zone: us-west-2b"))
-
-			By("Create cluster with both zone and subnet set")
-			out, err = clusterService.CreateDryRun(
-				clusterName, "--availability-zones", "us-west-2b", "--subnet-ids", "subnet-039f2a2a2d2d83e7f",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("Setting availability zones is not supported for BYO VPC. ROSA autodetects availability zones from subnet IDs provided"))
-		})
+				By("Create cluster with both zone and subnet set")
+				out, err = clusterService.CreateDryRun(
+					clusterName, "--availability-zones", "us-west-2b", "--subnet-ids", "subnet-039f2a2a2d2d83e7f",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("Setting availability zones is not supported for BYO VPC. ROSA autodetects availability zones from subnet IDs provided"))
+			})
 	})
 
 var _ = Describe("Classic cluster deletion validation",
-	labels.Day3,
-	labels.FeatureCluster,
+	labels.Feature.Cluster,
 	func() {
 		defer GinkgoRecover()
 
@@ -597,29 +606,30 @@ var _ = Describe("Classic cluster deletion validation",
 			rosaClient = rosacli.NewClient()
 		})
 
-		It("To validate the ROSA cluster deletion will work via rosacli	- [id:38778]", labels.Medium, labels.Day3, func() {
-			clusterService := rosaClient.Cluster
-			notExistID := "no-exist-cluster-id"
-			By("Delete the cluster without indicated cluster Name or ID")
-			cmd := []string{"rosa", "delete", "cluster"}
-			out, err := rosaClient.Runner.RunCMD(cmd)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("\"cluster\" not set"))
+		It("to validate the ROSA cluster deletion will work via rosacli	- [id:38778]",
+			labels.Medium, labels.Runtime.Day1Supplemental,
+			func() {
+				clusterService := rosaClient.Cluster
+				notExistID := "no-exist-cluster-id"
+				By("Delete the cluster without indicated cluster Name or ID")
+				cmd := []string{"rosa", "delete", "cluster"}
+				out, err := rosaClient.Runner.RunCMD(cmd)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("\"cluster\" not set"))
 
-			By("Delete a non-existed cluster")
-			out, err = clusterService.DeleteCluster(notExistID, "-y")
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("There is no cluster with identifier or name"))
+				By("Delete a non-existed cluster")
+				out, err = clusterService.DeleteCluster(notExistID, "-y")
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("There is no cluster with identifier or name"))
 
-			By("Delete with unknown flag --interactive")
-			out, err = clusterService.DeleteCluster(notExistID, "-y", "--interactive")
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("unknown flag: --interactive"))
-		})
+				By("Delete with unknown flag --interactive")
+				out, err = clusterService.DeleteCluster(notExistID, "-y", "--interactive")
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("unknown flag: --interactive"))
+			})
 	})
 var _ = Describe("Classic cluster creation negavite testing",
-	labels.Day1Negative,
-	labels.FeatureCluster,
+	labels.Feature.Cluster,
 	func() {
 		defer GinkgoRecover()
 
@@ -634,73 +644,74 @@ var _ = Describe("Classic cluster creation negavite testing",
 			clusterService = rosaClient.Cluster
 		})
 
-		It("to validate to create the sts cluster with the version not compatible with the role version	- [id:45176]", labels.Medium, func() {
-			clusterService = rosaClient.Cluster
-			ocmResourceService := rosaClient.OCMResource
+		It("to validate to create the sts cluster with the version not compatible with the role version	- [id:45176]",
+			labels.Medium, labels.Runtime.Day1Negative,
+			func() {
+				clusterService = rosaClient.Cluster
+				ocmResourceService := rosaClient.OCMResource
 
-			By("Porepare version for testing")
-			var accRoleversion string
-			versionService := rosaClient.Version
-			versionList, err := versionService.ListAndReflectVersions(rosacli.VersionChannelGroupStable, false)
-			Expect(err).To(BeNil())
-			defaultVersion := versionList.DefaultVersion()
-			Expect(defaultVersion).ToNot(BeNil())
-			lowerVersion, err := versionList.FindNearestBackwardMinorVersion(defaultVersion.Version, 1, true)
-			Expect(err).To(BeNil())
-			Expect(lowerVersion).NotTo(BeNil())
-
-			_, _, accRoleversion, err = lowerVersion.MajorMinor()
-			Expect(err).To(BeNil())
-
-			By("Create account-roles in low version 4.14")
-			accrolePrefix := "testAr45176"
-			path := "/a/b/"
-			output, err := ocmResourceService.CreateAccountRole("--mode", "auto",
-				"--prefix", accrolePrefix,
-				"--path", path,
-				"--version", accRoleversion,
-				"-y")
-			Expect(err).To(BeNil())
-			defer func() {
-				By("Delete the account-roles")
-				output, err := ocmResourceService.DeleteAccountRole("--mode", "auto",
-					"--prefix", accrolePrefix,
-					"-y")
-
+				By("Porepare version for testing")
+				var accRoleversion string
+				versionService := rosaClient.Version
+				versionList, err := versionService.ListAndReflectVersions(rosacli.VersionChannelGroupStable, false)
 				Expect(err).To(BeNil())
+				defaultVersion := versionList.DefaultVersion()
+				Expect(defaultVersion).ToNot(BeNil())
+				lowerVersion, err := versionList.FindNearestBackwardMinorVersion(defaultVersion.Version, 1, true)
+				Expect(err).To(BeNil())
+				Expect(lowerVersion).NotTo(BeNil())
+
+				_, _, accRoleversion, err = lowerVersion.MajorMinor()
+				Expect(err).To(BeNil())
+
+				By("Create account-roles in low version 4.14")
+				accrolePrefix := "testAr45176"
+				path := "/a/b/"
+				output, err := ocmResourceService.CreateAccountRole("--mode", "auto",
+					"--prefix", accrolePrefix,
+					"--path", path,
+					"--version", accRoleversion,
+					"-y")
+				Expect(err).To(BeNil())
+				defer func() {
+					By("Delete the account-roles")
+					output, err := ocmResourceService.DeleteAccountRole("--mode", "auto",
+						"--prefix", accrolePrefix,
+						"-y")
+
+					Expect(err).To(BeNil())
+					textData := rosaClient.Parser.TextData.Input(output).Parse().Tip()
+					Expect(textData).To(ContainSubstring("Successfully deleted"))
+				}()
 				textData := rosaClient.Parser.TextData.Input(output).Parse().Tip()
-				Expect(textData).To(ContainSubstring("Successfully deleted"))
-			}()
-			textData := rosaClient.Parser.TextData.Input(output).Parse().Tip()
-			Expect(textData).To(ContainSubstring("Created role"))
+				Expect(textData).To(ContainSubstring("Created role"))
 
-			arl, _, err := ocmResourceService.ListAccountRole()
-			Expect(err).To(BeNil())
-			ar := arl.DigAccountRoles(accrolePrefix, false)
-			fmt.Println(ar)
+				arl, _, err := ocmResourceService.ListAccountRole()
+				Expect(err).To(BeNil())
+				ar := arl.DigAccountRoles(accrolePrefix, false)
+				fmt.Println(ar)
 
-			By("Create cluster with latest version and use the low version account-roles")
-			clusterName := "cluster45176"
-			operatorRolePrefix := "cluster45176-xvfa"
-			out, err, _ := clusterService.Create(
-				clusterName, "--sts",
-				"--mode", "auto",
-				"--role-arn", ar.InstallerRole,
-				"--support-role-arn", ar.SupportRole,
-				"--controlplane-iam-role", ar.ControlPlaneRole,
-				"--worker-iam-role", ar.WorkerRole,
-				"--operator-roles-prefix", operatorRolePrefix,
-				"-y", "--dry-run",
-			)
-			Expect(err).NotTo(BeNil())
-			Expect(out.String()).To(ContainSubstring("is not compatible with version"))
-			Expect(out.String()).To(ContainSubstring("to create compatible roles and try again"))
-		})
+				By("Create cluster with latest version and use the low version account-roles")
+				clusterName := "cluster45176"
+				operatorRolePrefix := "cluster45176-xvfa"
+				out, err, _ := clusterService.Create(
+					clusterName, "--sts",
+					"--mode", "auto",
+					"--role-arn", ar.InstallerRole,
+					"--support-role-arn", ar.SupportRole,
+					"--controlplane-iam-role", ar.ControlPlaneRole,
+					"--worker-iam-role", ar.WorkerRole,
+					"--operator-roles-prefix", operatorRolePrefix,
+					"-y", "--dry-run",
+				)
+				Expect(err).NotTo(BeNil())
+				Expect(out.String()).To(ContainSubstring("is not compatible with version"))
+				Expect(out.String()).To(ContainSubstring("to create compatible roles and try again"))
+			})
 	})
 
 var _ = Describe("HCP cluster creation negative testing",
-	labels.Day1Negative,
-	labels.FeatureCluster,
+	labels.Feature.Cluster,
 	func() {
 		defer GinkgoRecover()
 
@@ -719,59 +730,61 @@ var _ = Describe("HCP cluster creation negative testing",
 			clusterService = rosaClient.Cluster
 		})
 
-		It("create HCP cluster with network type validation can work well via rosa cli - [id:73725]", labels.Medium, func() {
-			isHostedCP, err := clusterService.IsHostedCPCluster(clusterID)
-			Expect(err).To(BeNil())
+		It("create HCP cluster with network type validation can work well via rosa cli - [id:73725]",
+			labels.Medium, labels.Runtime.Day1Negative,
+			func() {
+				isHostedCP, err := clusterService.IsHostedCPCluster(clusterID)
+				Expect(err).To(BeNil())
 
-			rosalCommand, err := config.RetrieveClusterCreationCommand(ciConfig.Test.CreateCommandFile)
-			Expect(err).To(BeNil())
+				rosalCommand, err := config.RetrieveClusterCreationCommand(ciConfig.Test.CreateCommandFile)
+				Expect(err).To(BeNil())
 
-			if !isHostedCP {
-				By("Create non-HCP cluster with --no-cni flag")
-				clusterName := common.GenerateRandomName("classic-73725", 2)
-				operatorPrefix := common.GenerateRandomName("classic-oper", 2)
-				replacingFlags := map[string]string{
-					"-c":                     clusterName,
-					"--cluster-name":         clusterName,
-					"--domain-prefix":        clusterName,
-					"--operator-role-prefix": operatorPrefix,
+				if !isHostedCP {
+					By("Create non-HCP cluster with --no-cni flag")
+					clusterName := common.GenerateRandomName("classic-73725", 2)
+					operatorPrefix := common.GenerateRandomName("classic-oper", 2)
+					replacingFlags := map[string]string{
+						"-c":                     clusterName,
+						"--cluster-name":         clusterName,
+						"--domain-prefix":        clusterName,
+						"--operator-role-prefix": operatorPrefix,
+					}
+					rosalCommand.ReplaceFlagValue(replacingFlags)
+					rosalCommand.AddFlags("--dry-run", "--no-cni", "-y")
+					output, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+					Expect(err).To(HaveOccurred())
+					Expect(output.String()).To(ContainSubstring("ERR: Disabling CNI is supported only for Hosted Control Planes"))
+				} else {
+					By("Create HCP cluster with --no-cni and \"--network-type={OVNKubernetes, OpenshiftSDN}\" at the same time")
+					clusterName := common.GenerateRandomName("cluster-71946", 2)
+					operatorPrefix := common.GenerateRandomName("cluster-oper", 2)
+
+					replacingFlags := map[string]string{
+						"-c":                     clusterName,
+						"--cluster-name":         clusterName,
+						"--domain-prefix":        clusterName,
+						"--operator-role-prefix": operatorPrefix,
+					}
+					rosalCommand.ReplaceFlagValue(replacingFlags)
+					rosalCommand.AddFlags("--dry-run", "--no-cni", "--network-type='{OVNKubernetes,OpenshiftSDN}'", "-y")
+					output, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+					Expect(err).To(HaveOccurred())
+					Expect(output.String()).To(ContainSubstring("ERR: Expected a valid network type. Valid values: [OpenShiftSDN OVNKubernetes]"))
+
+					By("Create hcp cluster with invalid --no-cni value")
+					rosalCommand.DeleteFlag("--network-type", true)
+					rosalCommand.DeleteFlag("--no-cni", true)
+					rosalCommand.AddFlags("--no-cni=ui")
+					output, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+					Expect(err).To(HaveOccurred())
+					Expect(output.String()).To(ContainSubstring(`Failed to execute root command: invalid argument "ui" for "--no-cni" flag: strconv.ParseBool: parsing "ui": invalid syntax`))
+
+					By("Create hcp cluster with --no-cni and --network-type=OVNKubernetes at the same time")
+					rosalCommand.DeleteFlag("--no-cni=ui", false)
+					rosalCommand.AddFlags("--no-cni", "--network-type=OVNKubernetes")
+					output, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
+					Expect(err).To(HaveOccurred())
+					Expect(output.String()).To(ContainSubstring("ERR: --no-cni and --network-type are mutually exclusive parameters"))
 				}
-				rosalCommand.ReplaceFlagValue(replacingFlags)
-				rosalCommand.AddFlags("--dry-run", "--no-cni", "-y")
-				output, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-				Expect(err).To(HaveOccurred())
-				Expect(output.String()).To(ContainSubstring("ERR: Disabling CNI is supported only for Hosted Control Planes"))
-			} else {
-				By("Create HCP cluster with --no-cni and \"--network-type={OVNKubernetes, OpenshiftSDN}\" at the same time")
-				clusterName := common.GenerateRandomName("cluster-71946", 2)
-				operatorPrefix := common.GenerateRandomName("cluster-oper", 2)
-
-				replacingFlags := map[string]string{
-					"-c":                     clusterName,
-					"--cluster-name":         clusterName,
-					"--domain-prefix":        clusterName,
-					"--operator-role-prefix": operatorPrefix,
-				}
-				rosalCommand.ReplaceFlagValue(replacingFlags)
-				rosalCommand.AddFlags("--dry-run", "--no-cni", "--network-type='{OVNKubernetes,OpenshiftSDN}'", "-y")
-				output, err := rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-				Expect(err).To(HaveOccurred())
-				Expect(output.String()).To(ContainSubstring("ERR: Expected a valid network type. Valid values: [OpenShiftSDN OVNKubernetes]"))
-
-				By("Create hcp cluster with invalid --no-cni value")
-				rosalCommand.DeleteFlag("--network-type", true)
-				rosalCommand.DeleteFlag("--no-cni", true)
-				rosalCommand.AddFlags("--no-cni=ui")
-				output, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-				Expect(err).To(HaveOccurred())
-				Expect(output.String()).To(ContainSubstring(`Failed to execute root command: invalid argument "ui" for "--no-cni" flag: strconv.ParseBool: parsing "ui": invalid syntax`))
-
-				By("Create hcp cluster with --no-cni and --network-type=OVNKubernetes at the same time")
-				rosalCommand.DeleteFlag("--no-cni=ui", false)
-				rosalCommand.AddFlags("--no-cni", "--network-type=OVNKubernetes")
-				output, err = rosaClient.Runner.RunCMD(strings.Split(rosalCommand.GetFullCommand(), " "))
-				Expect(err).To(HaveOccurred())
-				Expect(output.String()).To(ContainSubstring("ERR: --no-cni and --network-type are mutually exclusive parameters"))
-			}
-		})
+			})
 	})
