+++ tests/e2e/test_rosacli_token.go
@@ -11,8 +11,7 @@ import (
 )
 
 var _ = Describe("rosacli token",
-	labels.Day1,
-	labels.FeatureToken,
+	labels.Feature.Token,
 	func() {
 		defer GinkgoRecover()
 
@@ -25,51 +24,53 @@ var _ = Describe("rosacli token",
 			rosaClient = rosacli.NewClient()
 		})
 
-		It("Generate token via `rosa token` command and check all flags of it - [id:72162]", labels.High, func() {
-			ocmService := rosaClient.OCMResource
+		It("Generate token via `rosa token` command and check all flags of it - [id:72162]",
+			labels.High, labels.Runtime.OCMResources,
+			func() {
+				ocmService := rosaClient.OCMResource
 
-			By("check `rosa token --header`")
-			parsedHeaderOutput, err := rosaClient.Runner.RunPipeline(
-				[]string{"rosa", "token"},
-				[]string{"cut", "-d", ".", "-f", "1"},
-				[]string{"tr", "_-", "/+"},
-				[]string{"base64", "-D"},
-			)
-			Expect(err).To(BeNil())
-			headerOutput, err := ocmService.Token("--header")
-			Expect(err).To(BeNil())
+				By("check `rosa token --header`")
+				parsedHeaderOutput, err := rosaClient.Runner.RunPipeline(
+					[]string{"rosa", "token"},
+					[]string{"cut", "-d", ".", "-f", "1"},
+					[]string{"tr", "_-", "/+"},
+					[]string{"base64", "-D"},
+				)
+				Expect(err).To(BeNil())
+				headerOutput, err := ocmService.Token("--header")
+				Expect(err).To(BeNil())
 
-			Expect(headerOutput.String()).To(ContainSubstring(parsedHeaderOutput.String()))
+				Expect(headerOutput.String()).To(ContainSubstring(parsedHeaderOutput.String()))
 
-			By("Generate new token")
-			originalToken, err := ocmService.Token()
-			Expect(err).To(BeNil())
-			newGeneratedToken, err := ocmService.Token("--generate")
-			Expect(err).To(BeNil())
-			Expect(originalToken.String()).ToNot(Equal(newGeneratedToken.String()))
+				By("Generate new token")
+				originalToken, err := ocmService.Token()
+				Expect(err).To(BeNil())
+				newGeneratedToken, err := ocmService.Token("--generate")
+				Expect(err).To(BeNil())
+				Expect(originalToken.String()).ToNot(Equal(newGeneratedToken.String()))
 
-			By("Check `rosa token --parload`")
-			parsedPayloadOutput, err := rosaClient.Runner.RunPipeline(
-				[]string{"rosa", "token"},
-				[]string{"cut", "-d", ".", "-f", "2"},
-				[]string{"tr", "_-", "/+"},
-				[]string{"base64", "-D"},
-			)
-			Expect(err).To(BeNil())
-			payloadOutput, err := ocmService.Token("--payload")
-			Expect(err).To(BeNil())
-			Expect(strings.TrimSpace(payloadOutput.String())).To(Equal(strings.TrimSpace(parsedPayloadOutput.String())))
+				By("Check `rosa token --parload`")
+				parsedPayloadOutput, err := rosaClient.Runner.RunPipeline(
+					[]string{"rosa", "token"},
+					[]string{"cut", "-d", ".", "-f", "2"},
+					[]string{"tr", "_-", "/+"},
+					[]string{"base64", "-D"},
+				)
+				Expect(err).To(BeNil())
+				payloadOutput, err := ocmService.Token("--payload")
+				Expect(err).To(BeNil())
+				Expect(strings.TrimSpace(payloadOutput.String())).To(Equal(strings.TrimSpace(parsedPayloadOutput.String())))
 
-			By("Checlk `rosa token --signature`")
-			parsedSignatureOutput, err := rosaClient.Runner.RunPipeline(
-				[]string{"rosa", "token"},
-				[]string{"cut", "-d", ".", "-f", "3"},
-				[]string{"tr", "_-", "/+"},
-				[]string{"base64", "-D"},
-			)
-			Expect(err).To(BeNil())
-			signatureOutput, err := ocmService.Token("--signature")
-			Expect(err).To(BeNil())
-			Expect(strings.TrimSpace(signatureOutput.String())).To(ContainSubstring(strings.TrimSpace(parsedSignatureOutput.String())))
-		})
+				By("Checlk `rosa token --signature`")
+				parsedSignatureOutput, err := rosaClient.Runner.RunPipeline(
+					[]string{"rosa", "token"},
+					[]string{"cut", "-d", ".", "-f", "3"},
+					[]string{"tr", "_-", "/+"},
+					[]string{"base64", "-D"},
+				)
+				Expect(err).To(BeNil())
+				signatureOutput, err := ocmService.Token("--signature")
+				Expect(err).To(BeNil())
+				Expect(strings.TrimSpace(signatureOutput.String())).To(ContainSubstring(strings.TrimSpace(parsedSignatureOutput.String())))
+			})
 	})
