+++ tests/e2e/utils.go
@@ -1,0 +1,21 @@
+package e2e
+
+import (
+	"fmt"
+
+	. "github.com/onsi/ginkgo/v2"
+)
+
+func SkipNotHosted() {
+	message := fmt.Sprintln("The test profile is not hosted")
+	Skip(message)
+}
+func SkipNotClassic() {
+	message := fmt.Sprintln("The test profile is not classic")
+	Skip(message)
+}
+
+func SkipTestOnFeature(feature string) {
+	message := fmt.Sprintf("The test profile is not configured with the feature: %s", feature)
+	Skip(message)
+}
